backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["shivvvuu <jhodshivam5122@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.2"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.5.0"
faker = "^37.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
